requirements.txt
altair==5.1.2
anyascii==0.3.2
astunparse==1.6.3
CacheControl==0.13.1
contourpy==1.2.0
contractions==0.1.73
firebase-admin==5.3.0
flatbuffers==23.5.26
fonttools==4.45.1
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio-status==1.59.3
httplib2==0.22.0
importlib-metadata==6.8.0
install==1.3.5
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
markdown-it-py==3.0.0
matplotlib==3.8.2
mdurl==0.1.2
mkl-service==2.4.0
ml-dtypes==0.2.0
msgpack==1.0.7
nltk==3.8.1
pandas==1.5.3
Pillow==10.1.0
proto-plus==1.22.3
protobuf==3.20.3
pyahocorasick==2.0.0
pyarrow==14.0.1
pyasn1-modules==0.2.8
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.31.0
regex==2023.10.3
requests-oauthlib==1.3.0
rich==13.7.0
rpds-py==0.13.1
scikit-learn==1.3.2
scipy==1.11.4
smmap==5.0.1
streamlit==1.28.2
streamlit-option-menu==0.3.6
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
textsearch==0.0.24
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
vaderSentiment==3.3.2
validators==0.22.0
watchdog==3.0.0
wordninja==2.0.0
zipp==3.17.0